requirements.txt
pandas
PyAutoGUI
python-dotenv
